refactor out pipelineable inference operator
