+++ merlin/systems/dag/ops/feast.py
@@ -7,7 +7,7 @@ from feast import FeatureStore, ValueType
 from merlin.core.protocols import Transformable
 from merlin.dag import ColumnSelector
 from merlin.schema import ColumnSchema, Schema
-from merlin.systems.dag.ops.operator import PipelineableInferenceOperator
+from merlin.systems.dag.ops.operator import InferenceOperator
 
 # Feast_key: (numpy dtype, is_list, is_ragged)
 feast_2_numpy = {
@@ -20,7 +20,7 @@ feast_2_numpy = {
 }
 
 
-class QueryFeast(PipelineableInferenceOperator):
+class QueryFeast(InferenceOperator):
     """
     The QueryFeast operator is responsible for ensuring that your feast feature store [1]
     can communicate correctly with tritonserver for the ensemble feast feature look ups.
