+++ merlin/systems/dag/ops/session_filter.py
@@ -21,10 +21,10 @@ import numpy as np
 from merlin.core.protocols import Transformable
 from merlin.dag import ColumnSelector, Node
 from merlin.schema import ColumnSchema, Schema
-from merlin.systems.dag.ops.operator import PipelineableInferenceOperator
+from merlin.systems.dag.ops.operator import InferenceOperator
 
 
-class FilterCandidates(PipelineableInferenceOperator):
+class FilterCandidates(InferenceOperator):
     """
     This operator takes the input column and filters out elements of that column
     based on the supplied criteria.
