+++ merlin/systems/dag/ops/unroll_features.py
@@ -21,10 +21,10 @@ from merlin.core.protocols import Transformable
 from merlin.dag import Node
 from merlin.dag.selector import ColumnSelector
 from merlin.schema import Schema
-from merlin.systems.dag.ops.operator import PipelineableInferenceOperator
+from merlin.systems.dag.ops.operator import InferenceOperator
 
 
-class UnrollFeatures(PipelineableInferenceOperator):
+class UnrollFeatures(InferenceOperator):
     """
     This operator takes a target column and joins the "unroll" columns to the target. This helps
     when broadcasting a series of user features to a set of items.
